Changed Due Date and Updated Game
Added functionality to take into account attackers armies located on the country they are attacking from. Added functionality to make a card trade if a player has 3 of the same cards. Added some more print statements to make the state of the game more clear during an attack. Changed the due date at the top of the main file.